+++ neoforge/src/main/java/com/verdantartifice/primalmagick/common/items/wands/ModularStaffItemNeoforge.java
@@ -1,5 +1,6 @@
 package com.verdantartifice.primalmagick.common.items.wands;
 
+import com.verdantartifice.primalmagick.common.wands.IWandComponent;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.UseOnContext;
@@ -23,5 +24,11 @@ public class ModularStaffItemNeoforge extends ModularStaffItem {
     public boolean canContinueUsing(ItemStack oldStack, ItemStack newStack) {
         // Don't break wand interaction just because the stack NBT changes
         return true;
+    }
+
+    @Override
+    public int getEnchantmentValue(ItemStack stack) {
+        // The enchantability of a staff is determined by its components
+        return this.getComponents(stack).stream().mapToInt(IWandComponent::getEnchantability).sum();
     }
 }